labs/40-AIAgents/a2a/requirements.txt
python-dotenv
httpx
azure-identity
uvicorn
starlette
sse-starlette
fastapi
a2a-sdk
